pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latteries"
version = "0.1.5"
description = "James' API LLM evaluations workflow library - A collection of tools for LLM API calls, caching, and evaluation workflows"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "James", email = "your-email@example.com" }
]
keywords = ["llm", "api", "evaluation", "openai", "anthropic", "caching"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "slist",
    "streamlit",
    "anyio",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "httpx",
    "tqdm",
    "python-dotenv",
    "tinker>=0.5.1",
    "plotly",
    "transformers",
    "termcolor",
    "rich",
    "numpy",
    "chz",
    "wandb",
    "datasets",
    "tiktoken"
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pyright",
    "pytest",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/thejaminator/latteries"
Repository = "https://github.com/thejaminator/latteries"
Issues = "https://github.com/thejaminator/latteries/issues"

[project.scripts]
latteries-viewer = "latteries.viewer:main"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.black]
line-length = 120
exclude = "venv*"
target-version = ['py311']

[tool.ruff]
line-length = 120

[tool.pyright]
exclude = [".venv"]
venvPath = "."
venv = ".venv"
